pyproject.toml
[project]
name = "mcp-server-datahub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "acryl-datahub>=0.15.0.5",
    "mcp[cli]>=1.3.0",
]

[tool.uv.sources]
acryl-datahub = { path = "../datahub/metadata-ingestion", editable = true }
